requirements.txt
amqp==5.3.1
arrow==1.3.0
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
blessed==1.20.0
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==44.0.2
daphne==4.1.2
Django==5.1.6
django-cors-headers==4.7.0
django-crontab==0.7.1
django-picklefield==3.2
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kombu==5.4.2
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
prompt_toolkit==3.0.50
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.0.0
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==3.5.3
reportlab==4.3.1
requests==2.32.3
rest-framework-simplejwt==0.0.2
service-identity==24.2.0
setuptools==75.8.2
six==1.17.0
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
zope.interface==7.2
